Add left join test
